Refactor swiftype extensions into generic instance
- The extensions for File and SiteTree had the essential the same functionality with one or two lines of difference. The main functionality was pulled into new class SwiftTypeCrawlerExtension
- New class provides new methods to get object links and exclude records from indexing
